updated api, now only one memo create endpoint
